requirements.txt
cassandra-driver
yfinance
fredapi
python-dotenv
happybase
pyspark
